fix: fix unit test
